requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-storage-blob==12.20.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
joblib==1.4.2
nltk==3.8.1
numpy==2.0.0
openai==1.35.10
pandas==2.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
